# Vectorize `dotProduct`