--- a/src/Functions/array/arrayDotProduct.cpp
+++ b/src/Functions/array/arrayDotProduct.cpp
@@ -130,7 +130,7 @@ struct DotProduct
 
 
 /// The implementation is modeled after the implementation of distance functions arrayL1Distance, arrayL2Distance, etc.
-/// The main difference is that arrayDotProduct() interfers the result type differently.
+/// The main difference is that arrayDotProduct() interferes the result type differently.
 template <typename Kernel>
 class FunctionArrayScalarProduct : public IFunction
 {