--- a/tests/queries/0_stateless/02708_dotProduct.reference
+++ b/tests/queries/0_stateless/02708_dotProduct.reference
@@ -1,0 +1,34 @@
+-- Negative tests
+-- Tests
+   -- Array
+[1,2,3]	[4,5,6]	32	UInt16
+[1,2,3]	[4,5,6]	32	UInt32
+[1,2,3]	[4,5,6]	32	UInt64
+[1,2,3]	[4,5,6]	32	UInt64
+[-1,-2,-3]	[4,5,6]	-32	Int16
+[-1,-2,-3]	[4,5,6]	-32	Int32
+[-1,-2,-3]	[4,5,6]	-32	Int64
+[-1,-2,-3]	[4,5,6]	-32	Int64
+[1,2,3]	[4,5,6]	32	Float32
+[1,2,3]	[4,5,6]	32	Float64
+   -- Tuple
+(1,2,3)	(4,5,6)	32	UInt64
+(1,2,3)	(4,5,6)	32	UInt64
+(1,2,3)	(4,5,6)	32	UInt64
+(1,2,3)	(4,5,6)	32	UInt64
+(-1,-2,-3)	(4,5,6)	-32	Int64
+(-1,-2,-3)	(4,5,6)	-32	Int64
+(-1,-2,-3)	(4,5,6)	-32	Int64
+(-1,-2,-3)	(4,5,6)	-32	Int64
+(1,2,3)	(4,5,6)	32	Float64
+(1,2,3)	(4,5,6)	32	Float64
+-- Non-const argument
+[1,2,3]	[4,5,6]	32	UInt16
+ -- Array with mixed element arguments types (result type is the supertype)
+[1,2,3]	[4,5,6]	32	Float32
+ -- Tuple with mixed element arguments types
+(1,2,3)	(4,5,6)	32	Float64
+-- Aliases
+32
+32
+32