… docs/en/sql-reference/functions/tuple-functions.md
… tests/performance/dotProduct.xml
… tests/queries/0_stateless/02708_dotProduct.sql
… tests/queries/0_stateless/02708_dot_product.reference
… tests/queries/0_stateless/02708_dot_product.sql