Fixes #58096

This PR is similar in spirit as #58866.

Review note. The PR includes these commits:
- (please ignore) Refactoring and cleanups: https://github.com/ClickHouse/ClickHouse/commit/0fff496ad388df450af538df2cc3a57c29c86f45
- (please ignore) More refactorings: https://github.com/ClickHouse/ClickHouse/commit/db974098d073496d640bf08f0e43f3aa572ad3ff
- A try to extend the existing tests and make them more systematic: https://github.com/ClickHouse/ClickHouse/commit/877dc695f29b0df163a847ad323eaf2d20552af5
- New performance tests for `dotProduct`: https://github.com/ClickHouse/ClickHouse/commit/7d354164a5c47b762cfac2aea7406cd99fc00fe0
- Some help for the compiler with auto-vectorization (*): https://github.com/ClickHouse/ClickHouse/commit/e57b8f64e0e8112c063daacba4933145000603de
- Manual vectorization for `Float32/Float64` cases (**): https://github.com/ClickHouse/ClickHouse/commit/daa61a8576918ef650e314d06f4ece08cbbc74e7
- (please ignore) Further refactorings to make the code more similar to arrayDistance.cpp: https://github.com/ClickHouse/ClickHouse/pull/60202/commits/9f37f179005089c58f8f7cdacc081e9b203bcf38

(*) For query
   `SELECT sum(dp) FROM (SELECT dotProduct(v, v) AS dp FROM vectors)`
with vectors of type
   `Array(Float32)` and
   `Array(Float64)`
and dimension = 150, better auto-vectorization reduces runtimes from
  0.46 sec to 0.34 sec (`Float32`) and
  0.74 sec to 0.68 sec (`Float64`))

(**) Manual vectorization further improves runtimes from
  0.34 sec to 0.30 sec (`Float32`) and
  0.68 sec to 0.54 sec (`Float64`))

Improvements are not revolutionary but in the same ballpark as in #58866.

@gingerwizard 

### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Vectorized function `dotProduct` which is useful for vector search